refactor: using cryptos asData to update moves and coinBalanceProvider
