+++ lib/moves/view/moves_screen.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
-import 'package:projeto_criptos/moves/controller/movements_provider.dart';
+import 'package:projeto_criptos/shared/user/movements_provider.dart';
 import 'package:projeto_criptos/shared/templates/app_assets.dart';
 
 import '../widgets/movements_list_tile.dart';
