+++ lib/review/review_page.dart
@@ -1,19 +1,17 @@
 import 'package:decimal/decimal.dart';
 import 'package:flutter/material.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
-import 'package:projeto_criptos/moves/controller/movements_provider.dart';
-import 'package:projeto_criptos/moves/model/movement_model.dart';
+import 'package:projeto_criptos/shared/user/movements_provider.dart';
 import 'package:projeto_criptos/portfolio/controller/cryptos_provider.dart';
-import 'package:projeto_criptos/review/model/moves_model.dart';
 import 'package:projeto_criptos/review/widgets/info_review_column.dart';
 
 import 'package:projeto_criptos/review/widgets/review_buttons.dart';
 import 'package:projeto_criptos/shared/common_model/crypto.dart';
-import 'package:projeto_criptos/shared/templates/error_body.dart';
-import 'package:projeto_criptos/shared/templates/loading_body.dart';
 import 'package:projeto_criptos/shared/templates/model_app_bar.dart';
 import 'package:projeto_criptos/shared/user/user_coin_ammount_provider.dart';
 import 'package:projeto_criptos/shared/utils/decimal_to_double.dart';
+
+import '../shared/common_model/move_model.dart';
 
 class RevisionPage extends ConsumerStatefulWidget {
   const RevisionPage({
@@ -40,12 +38,12 @@ class _$RevisionPageState extends ConsumerState<RevisionPage> {
     return '1 ${widget.convertCoin.symbol.toUpperCase()} = ${exchange.toStringAsFixed(7)} ${widget.recieveCoin.symbol.toUpperCase()}';
   }
 
-  late MovesModel moves;
+  late MoveModel moves;
 
   @override
   Widget build(BuildContext context) {
     final cryptos = ref.watch(cryptosProvider);
-    moves = MovesModel(
+    moves = MoveModel(
       convert: widget.convert,
       recieve: widget.recieve,
       data: DateTime.now(),
@@ -53,65 +51,52 @@ class _$RevisionPageState extends ConsumerState<RevisionPage> {
       convertCoinId: widget.convertCoin.symbol,
       recieveCoinId: widget.recieveCoin.symbol,
     );
-    return cryptos.when(data: (data) {
-      return Scaffold(
-        appBar: const ModelAppBar(text: 'Revisar'),
-        body: Padding(
-          padding: const EdgeInsets.symmetric(
-            vertical: 25,
-            horizontal: 30,
-          ),
-          child: Column(
-            crossAxisAlignment: CrossAxisAlignment.stretch,
-            children: [
-              Text(
-                'Revise os dados da sua conversão',
-                style: Theme.of(context).textTheme.headlineLarge,
-              ),
-              InfoReviewColumn(
-                convert:
-                    '${widget.convert.toString()} ${widget.convertCoin.symbol.toUpperCase()}',
-                recieve:
-                    '${widget.recieve.toStringAsFixed(6)} ${widget.recieveCoin.symbol.toUpperCase()}',
-                cambio: getExchange(),
-              ),
-              ReviewButtons(
-                onPressed: () {
-                  int convertedId = data.indexWhere(
-                      (crypto) => crypto.id == widget.convertCoin.id);
-                  int recieveId = data.indexWhere(
-                      (crypto) => crypto.id == widget.recieveCoin.id);
-                  setState(() {
-                    ref
-                        .read(userCoinAmmountProvider.notifier)
-                        .state[convertedId] -= dtd(widget.convert);
-                    ref
-                        .read(userCoinAmmountProvider.notifier)
-                        .state[recieveId] += dtd(widget.recieve);
-                    ref
-                        .read(movementsProvider.state)
-                        .state
-                        .add(MovementModel.fromMap(moves.toMap()));
-                  });
-                  Navigator.pushReplacementNamed(
-                    context,
-                    '/success',
-                    arguments: moves,
-                  );
-                },
-              ),
-            ],
-          ),
+    return Scaffold(
+      appBar: const ModelAppBar(text: 'Revisar'),
+      body: Padding(
+        padding: const EdgeInsets.symmetric(
+          vertical: 25,
+          horizontal: 30,
         ),
-      );
-    }, error: (e, r) {
-      return ErrorBody(
-        onError: () {
-          ref.refresh(cryptosProvider);
-        },
-      );
-    }, loading: () {
-      return const LoadingBody();
-    });
+        child: Column(
+          crossAxisAlignment: CrossAxisAlignment.stretch,
+          children: [
+            Text(
+              'Revise os dados da sua conversão',
+              style: Theme.of(context).textTheme.headlineLarge,
+            ),
+            InfoReviewColumn(
+              convert:
+                  '${widget.convert.toString()} ${widget.convertCoin.symbol.toUpperCase()}',
+              recieve:
+                  '${widget.recieve.toStringAsFixed(6)} ${widget.recieveCoin.symbol.toUpperCase()}',
+              cambio: getExchange(),
+            ),
+            ReviewButtons(
+              onPressed: () {
+                int convertedId = cryptos.asData!.value
+                    .indexWhere((crypto) => crypto.id == widget.convertCoin.id);
+
+                int recieveId = cryptos.asData!.value
+                    .indexWhere((crypto) => crypto.id == widget.recieveCoin.id);
+                setState(() {
+                  ref
+                      .read(userCoinAmmountProvider.notifier)
+                      .state[convertedId] -= dtd(widget.convert);
+                  ref.read(userCoinAmmountProvider.notifier).state[recieveId] +=
+                      dtd(widget.recieve);
+                  ref.read(movementsProvider.state).state.add(moves);
+                });
+                Navigator.pushReplacementNamed(
+                  context,
+                  '/success',
+                  arguments: moves,
+                );
+              },
+            ),
+          ],
+        ),
+      ),
+    );
   }
 }
