+++ lib/review/success_conversion_page.dart
@@ -2,10 +2,8 @@ import 'dart:math';
 
 import 'package:auto_size_text/auto_size_text.dart';
 import 'package:flutter/material.dart';
-import 'package:projeto_criptos/shared/common_model/move.dart';
+import 'package:projeto_criptos/shared/common_model/move_model.dart';
 import 'package:projeto_criptos/shared/templates/app_assets.dart';
-
-import 'model/moves_model.dart';
 
 class SucessConversionPage extends StatefulWidget {
   const SucessConversionPage({
