+++ lib/route_controller.dart
@@ -1,8 +1,7 @@
 import 'package:flutter/material.dart';
-import 'package:projeto_criptos/review/model/moves_model.dart';
 import 'package:projeto_criptos/review/success_conversion_page.dart';
 import 'package:projeto_criptos/review/review_page.dart';
-import 'package:projeto_criptos/shared/common_model/move.dart';
+import 'package:projeto_criptos/shared/common_model/move_model.dart';
 import 'package:projeto_criptos/shared/utils/arguments/to_revision_arguments.dart';
 import 'conversion/view/conversion_page.dart';
 import 'shared/utils/arguments/to_conversion_arguments.dart';
@@ -63,7 +62,7 @@ class RouteController {
         },
       );
     } else if (settings.name == SucessConversionPage.route) {
-      final args = settings.arguments as MovesModel;
+      final args = settings.arguments as MoveModel;
       return PageRouteBuilder(
         settings: settings,
         pageBuilder: (context, animation, secondaryAnimation) {
