+++ lib/shared/user/movements_provider.dart
@@ -1,11 +1,12 @@
 import 'package:hooks_riverpod/hooks_riverpod.dart';
-import 'package:projeto_criptos/moves/model/movement_model.dart';
 import 'package:projeto_criptos/shared/utils/decimal_parse.dart';
+
+import '../common_model/move_model.dart';
 
 final movementsProvider = StateProvider(
   (ref) {
     return [
-      MovementModel(
+      MoveModel(
         convert: dp('5.21'),
         recieve: dp('2.21'),
         data: DateTime.now(),
